Refactor OutlinePreviewCanvas to use ImageOutlinePass for rendering
