+++ include/ui/OutlinePreviewCanvas.h
@@ -9,6 +9,7 @@ class SceneManager;
 class SoSeparator;
 class SoCamera;
 class SoPerspectiveCamera;
+class ImageOutlinePass;
 
 class OutlinePreviewCanvas : public wxGLCanvas {
 public:
@@ -45,39 +46,15 @@ private:
     void render();
     int getObjectAtPosition(const wxPoint& pos);
     
-    // FBO and shader management
-    void initializeFBO(int width, int height);
-    void cleanupFBO();
-    void initializeShaders();
-    void cleanupShaders();
-    unsigned int compileShader(const char* source, unsigned int type);
-    unsigned int createShaderProgram(const char* vertexSource, const char* fragmentSource);
-    void createQuadVAO();
-    
     wxGLContext* m_glContext{ nullptr };
     std::unique_ptr<SceneManager> m_sceneManager;
+    std::unique_ptr<ImageOutlinePass> m_outlinePass;
     SoSeparator* m_sceneRoot{ nullptr };
     SoSeparator* m_modelRoot{ nullptr };
     SoCamera* m_camera{ nullptr };
     
     ImageOutlineParams m_outlineParams;
     bool m_outlineEnabled{ true };
-    
-    // FBO and texture resources
-    unsigned int m_fbo{ 0 };
-    unsigned int m_colorTexture{ 0 };
-    unsigned int m_depthTexture{ 0 };
-    unsigned int m_normalTexture{ 0 };
-    
-    // Shader programs
-    unsigned int m_normalShader{ 0 };
-    unsigned int m_outlineShader{ 0 };
-    unsigned int m_quadVAO{ 0 };
-    unsigned int m_quadVBO{ 0 };
-    
-    // FBO size
-    int m_fboWidth{ 0 };
-    int m_fboHeight{ 0 };
     
     bool m_initialized{ false };
     bool m_needsRedraw{ true };
